backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
ccxt==4.1.10
python-dotenv==1.0.0
websockets==12.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1